+++ package.json
@@ -8,7 +8,7 @@
     "lint:fix": "prettier --write '{src,test}/**/*' README.md package.json",
     "pretest": "npm run -s lint",
     "test": "jest --coverage",
-    "test:typescript": "npx tsc --noEmit --declaration --noUnusedLocals test/typescript-validate.ts"
+    "test:typescript": "npx tsc --noEmit --declaration --noUnusedLocals --allowImportingTsExtensions --moduleResolution node16 --module node16 test/typescript-validate.ts"
   },
   "repository": "github:octokit/oauth-app.js",
   "keywords": [
@@ -60,6 +60,10 @@
         "functions": 100,
         "lines": 100
       }
+    },
+    "moduleNameMapper": {
+      "ipaddr.js": "<rootDir>/node_modules/ipaddr.js/lib/ipaddr.js",
+      "^(.+)\\.jsx?$": "$1"
     }
   },
   "release": {
