+++ src/methods/get-oauth-client-code.ts
@@ -1,5 +1,5 @@
 export function getOAuthClientCode() {
-  return `import { Octokit: Core } from "https://cdn.pika.dev/@octokit/core";
+  return `import { Octokit: Core } from "https://esm.sh/@octokit/core";
     
     export const Octokit = Core.defaults({
       oauth: {}
