+++ src/methods/get-web-flow-authorization-url.ts
@@ -1,6 +1,6 @@
 import * as OAuthMethods from "@octokit/oauth-methods";
 
-import type { ClientType, State } from "../types";
+import type { ClientType, State } from "../types.js";
 
 type StateOptions = "clientType" | "clientId" | "clientSecret" | "request";
