+++ src/methods/refresh-token.ts
@@ -1,7 +1,7 @@
 import * as OAuthMethods from "@octokit/oauth-methods";
 
-import type { State } from "../types";
-import { emitEvent } from "../emit-event";
+import type { State } from "../types.js";
+import { emitEvent } from "../emit-event.js";
 import { createOAuthUserAuth } from "@octokit/auth-oauth-user";
 
 export type RefreshTokenOptions = {
