+++ src/methods/scope-token.ts
@@ -1,8 +1,8 @@
 import * as OAuthMethods from "@octokit/oauth-methods";
 import { createOAuthUserAuth } from "@octokit/auth-oauth-user";
 
-import type { State } from "../types";
-import { emitEvent } from "../emit-event";
+import type { State } from "../types.js";
+import { emitEvent } from "../emit-event.js";
 
 type StateOptions = "clientType" | "clientId" | "clientSecret" | "request";
