+++ src/middleware/aws-lambda/api-gateway-v2-send-response.ts
@@ -1,4 +1,4 @@
-import type { OctokitResponse } from "../types";
+import type { OctokitResponse } from "../types.js";
 import type { APIGatewayProxyStructuredResultV2 } from "aws-lambda";
 
 export function sendResponse(
