+++ src/middleware/node/send-response.ts
@@ -2,7 +2,7 @@
 // see https://github.com/octokit/octokit.js/issues/2075#issuecomment-817361886
 // import { IncomingMessage, ServerResponse } from "node:http";
 type ServerResponse = any;
-import type { OctokitResponse } from "../types";
+import type { OctokitResponse } from "../types.js";
 
 export function sendResponse(
   octokitResponse: OctokitResponse,
