+++ src/oauth-app-octokit.ts
@@ -1,6 +1,6 @@
 import { Octokit } from "@octokit/core";
 import { getUserAgent } from "universal-user-agent";
-import { VERSION } from "./version";
+import { VERSION } from "./version.js";
 
 export const OAuthAppOctokit = Octokit.defaults({
   userAgent: `octokit-oauth-app.js/${VERSION} ${getUserAgent()}`,
